+++ packages/utils/src/misc.ts
@@ -1,8 +1,6 @@
-/// <reference lib="dom" />
+import { Event, Integration, WrappedFunction } from "@sentry/types";
 
-import { Event, Integration, WrappedFunction } from '@sentry/types';
-
-import { isString } from './is';
+import { isString } from "./is";
 
 /** Internal */
 interface SentryGlobal {
@@ -38,7 +36,11 @@ export function dynamicRequire(mod: any, request: string): any {
  */
 export function isNodeEnv(): boolean {
   // tslint:disable:strict-type-predicates
-  return Object.prototype.toString.call(typeof process !== 'undefined' ? process : 0) === '[object process]';
+  return (
+    Object.prototype.toString.call(
+      typeof process !== "undefined" ? process : 0
+    ) === "[object process]"
+  );
 }
 
 const fallbackGlobalObject = {};
@@ -51,9 +53,9 @@ const fallbackGlobalObject = {};
 export function getGlobalObject<T>(): T & SentryGlobal {
   return (isNodeEnv()
     ? global
-    : typeof window !== 'undefined'
+    : typeof window !== "undefined"
     ? window
-    : typeof self !== 'undefined'
+    : typeof self !== "undefined"
     ? self
     : fallbackGlobalObject) as T & SentryGlobal;
 }
@@ -96,15 +98,22 @@ export function uuid4(): string {
     };
 
     return (
-      pad(arr[0]) + pad(arr[1]) + pad(arr[2]) + pad(arr[3]) + pad(arr[4]) + pad(arr[5]) + pad(arr[6]) + pad(arr[7])
+      pad(arr[0]) +
+      pad(arr[1]) +
+      pad(arr[2]) +
+      pad(arr[3]) +
+      pad(arr[4]) +
+      pad(arr[5]) +
+      pad(arr[6]) +
+      pad(arr[7])
     );
   }
   // http://stackoverflow.com/questions/105034/how-to-create-a-guid-uuid-in-javascript/2117523#2117523
-  return 'xxxxxxxxxxxx4xxxyxxxxxxxxxxxxxxx'.replace(/[xy]/g, c => {
+  return "xxxxxxxxxxxx4xxxyxxxxxxxxxxxxxxx".replace(/[xy]/g, c => {
     // tslint:disable-next-line:no-bitwise
     const r = (Math.random() * 16) | 0;
     // tslint:disable-next-line:no-bitwise
-    const v = c === 'x' ? r : (r & 0x3) | 0x8;
+    const v = c === "x" ? r : (r & 0x3) | 0x8;
     return v.toString(16);
   });
 }
@@ -117,7 +126,7 @@ export function uuid4(): string {
  * @returns parsed URL object
  */
 export function parseUrl(
-  url: string,
+  url: string
 ): {
   host?: string;
   path?: string;
@@ -128,20 +137,22 @@ export function parseUrl(
     return {};
   }
 
-  const match = url.match(/^(([^:\/?#]+):)?(\/\/([^\/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?$/);
+  const match = url.match(
+    /^(([^:\/?#]+):)?(\/\/([^\/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?$/
+  );
 
   if (!match) {
     return {};
   }
 
   // coerce to undefined values to empty string so we don't get 'undefined'
-  const query = match[6] || '';
-  const fragment = match[8] || '';
+  const query = match[6] || "";
+  const fragment = match[8] || "";
   return {
     host: match[4],
     path: match[5],
     protocol: match[2],
-    relative: match[5] + query + fragment, // everything minus origin
+    relative: match[5] + query + fragment // everything minus origin
   };
 }
 
@@ -159,9 +170,9 @@ export function getEventDescription(event: Event): string {
     if (exception.type && exception.value) {
       return `${exception.type}: ${exception.value}`;
     }
-    return exception.type || exception.value || event.event_id || '<unknown>';
+    return exception.type || exception.value || event.event_id || "<unknown>";
   }
-  return event.event_id || '<unknown>';
+  return event.event_id || "<unknown>";
 }
 
 /** JSDoc */
@@ -172,9 +183,9 @@ interface ExtensibleConsole extends Console {
 /** JSDoc */
 export function consoleSandbox(callback: () => any): any {
   const global = getGlobalObject<Window>();
-  const levels = ['debug', 'info', 'warn', 'error', 'log', 'assert'];
+  const levels = ["debug", "info", "warn", "error", "log", "assert"];
 
-  if (!('console' in global)) {
+  if (!("console" in global)) {
     return callback();
   }
 
@@ -183,9 +194,16 @@ export function consoleSandbox(callback: () => any): any {
 
   // Restore all wrapped console methods
   levels.forEach(level => {
-    if (level in global.console && (originalConsole[level] as WrappedFunction).__sentry__) {
-      wrappedLevels[level] = (originalConsole[level] as WrappedFunction).__sentry_wrapped__;
-      originalConsole[level] = (originalConsole[level] as WrappedFunction).__sentry_original__;
+    if (
+      level in global.console &&
+      (originalConsole[level] as WrappedFunction).__sentry__
+    ) {
+      wrappedLevels[level] = (originalConsole[
+        level
+      ] as WrappedFunction).__sentry_wrapped__;
+      originalConsole[level] = (originalConsole[
+        level
+      ] as WrappedFunction).__sentry_original__;
     }
   });
 
@@ -207,12 +225,18 @@ export function consoleSandbox(callback: () => any): any {
  * @param type Type of the exception.
  * @hidden
  */
-export function addExceptionTypeValue(event: Event, value?: string, type?: string): void {
+export function addExceptionTypeValue(
+  event: Event,
+  value?: string,
+  type?: string
+): void {
   event.exception = event.exception || {};
   event.exception.values = event.exception.values || [];
   event.exception.values[0] = event.exception.values[0] || {};
-  event.exception.values[0].value = event.exception.values[0].value || value || '';
-  event.exception.values[0].type = event.exception.values[0].type || type || 'Error';
+  event.exception.values[0].value =
+    event.exception.values[0].value || value || "";
+  event.exception.values[0].type =
+    event.exception.values[0].type || type || "Error";
 }
 
 /**
@@ -225,13 +249,14 @@ export function addExceptionMechanism(
   event: Event,
   mechanism: {
     [key: string]: any;
-  } = {},
+  } = {}
 ): void {
   // TODO: Use real type with `keyof Mechanism` thingy and maybe make it better?
   try {
     // @ts-ignore
     // tslint:disable:no-non-null-assertion
-    event.exception!.values![0].mechanism = event.exception!.values![0].mechanism || {};
+    event.exception!.values![0].mechanism =
+      event.exception!.values![0].mechanism || {};
     Object.keys(mechanism).forEach(key => {
       // @ts-ignore
       event.exception!.values![0].mechanism[key] = mechanism[key];
@@ -248,7 +273,7 @@ export function getLocationHref(): string {
   try {
     return document.location.href;
   } catch (oO) {
-    return '';
+    return "";
   }
 }
 
@@ -258,29 +283,37 @@ export function getLocationHref(): string {
  * e.g. [HTMLElement] => body > div > input#foo.btn[name=baz]
  * @returns generated DOM path
  */
-export function htmlTreeAsString(elem: Node): string {
+export function htmlTreeAsString(elem: unknown): string {
+  type SimpleNode = {
+    parentNode: SimpleNode;
+  } | null;
+
   // try/catch both:
   // - accessing event.target (see getsentry/raven-js#838, #768)
   // - `htmlTreeAsString` because it's complex, and just accessing the DOM incorrectly
   // - can throw an exception in some circumstances.
   try {
-    let currentElem: Node | null = elem;
+    let currentElem = elem as SimpleNode;
     const MAX_TRAVERSE_HEIGHT = 5;
     const MAX_OUTPUT_LEN = 80;
     const out = [];
     let height = 0;
     let len = 0;
-    const separator = ' > ';
+    const separator = " > ";
     const sepLength = separator.length;
     let nextStr;
 
     while (currentElem && height++ < MAX_TRAVERSE_HEIGHT) {
-      nextStr = _htmlElementAsString(currentElem as HTMLElement);
+      nextStr = _htmlElementAsString(currentElem);
       // bail out if
       // - nextStr is the 'html' element
       // - the length of the string that would be created exceeds MAX_OUTPUT_LEN
       //   (ignore this limit if we are on the first iteration)
-      if (nextStr === 'html' || (height > 1 && len + out.length * sepLength + nextStr.length >= MAX_OUTPUT_LEN)) {
+      if (
+        nextStr === "html" ||
+        (height > 1 &&
+          len + out.length * sepLength + nextStr.length >= MAX_OUTPUT_LEN)
+      ) {
         break;
       }
 
@@ -292,7 +325,7 @@ export function htmlTreeAsString(elem: Node): string {
 
     return out.reverse().join(separator);
   } catch (_oO) {
-    return '<unknown>';
+    return "<unknown>";
   }
 }
 
@@ -301,7 +334,14 @@ export function htmlTreeAsString(elem: Node): string {
  * e.g. [HTMLElement] => input#foo.btn[name=baz]
  * @returns generated DOM path
  */
-function _htmlElementAsString(elem: HTMLElement): string {
+function _htmlElementAsString(el: unknown): string {
+  const elem = el as {
+    getAttribute(key: string): string; // tslint:disable-line:completed-docs
+    tagName?: string;
+    id?: string;
+    className?: string;
+  };
+
   const out = [];
   let className;
   let classes;
@@ -310,7 +350,7 @@ function _htmlElementAsString(elem: HTMLElement): string {
   let i;
 
   if (!elem || !elem.tagName) {
-    return '';
+    return "";
   }
 
   out.push(elem.tagName.toLowerCase());
@@ -325,7 +365,7 @@ function _htmlElementAsString(elem: HTMLElement): string {
       out.push(`.${classes[i]}`);
     }
   }
-  const attrWhitelist = ['type', 'name', 'title', 'alt'];
+  const attrWhitelist = ["type", "name", "title", "alt"];
   for (i = 0; i < attrWhitelist.length; i++) {
     key = attrWhitelist[i];
     attr = elem.getAttribute(key);
@@ -333,7 +373,7 @@ function _htmlElementAsString(elem: HTMLElement): string {
       out.push(`[${key}="${attr}"]`);
     }
   }
-  return out.join('');
+  return out.join("");
 }
 
 /**
@@ -371,6 +411,34 @@ export function parseSemver(input: string): SemVer {
     major: isNaN(major) ? undefined : major,
     minor: isNaN(minor) ? undefined : minor,
     patch: isNaN(patch) ? undefined : patch,
-    prerelease: match[4],
+    prerelease: match[4]
   };
+}
+
+const defaultRetryAfter = 60 * 1000; // 60 seconds
+
+/**
+ * Extracts Retry-After value from the request header or returns default value
+ * @param now current unix timestamp
+ * @param header string representation of 'Retry-After' header
+ */
+export function parseRetryAfterHeader(
+  now: number,
+  header?: string | number | null
+): number {
+  if (!header) {
+    return defaultRetryAfter;
+  }
+
+  const headerDelay = parseInt(`${header}`, 10);
+  if (!isNaN(headerDelay)) {
+    return headerDelay * 1000;
+  }
+
+  const headerDate = Date.parse(`${header}`);
+  if (!isNaN(headerDate)) {
+    return headerDate - now;
+  }
+
+  return defaultRetryAfter;
 }