+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 - [hub] feat: Add more span functions (#2161)
 - [integrations] feat: Change `Tracing` integration (#2161)
 
+## 5.9.0
+
+- [node] feat: Added `mode` option for `OnUnhandledRejection` integration that changes how we log errors and what we do with the process itself
+- [browser] ref: Both global handlers now always return `true` to call default implementations (error logging)
+
+## 5.8.0
+
+- [browser/node] feat: 429 http code handling in node/browser transports (#2300)
+- [core] feat: Make sure that Debug integration is always setup as the last one (#2285)
+- [browser] fix: Gracefuly handle incorrect input from onerror (#2302)
+- [utils] fix: Safer normalizing for input with `domain` key (#2305)
+- [utils] ref: Remove dom references from utils for old TS and env interop (#2303)
+
 ## 5.7.1
 
 - [core] ref: Use the smallest possible interface for our needs - `PromiseLike` (#2273)